+++ src/main/kotlin/dev/m2en/citation/Main.kt
@@ -17,6 +17,8 @@ import dev.kord.core.event.message.ReactionAddEvent
 import dev.kord.gateway.PrivilegedIntent
 import dev.m2en.citation.command.*
 import dev.m2en.citation.command.chat.HelpCommand
+import dev.m2en.citation.command.message.onDebugMessageCommand
+import dev.m2en.citation.command.message.onRegister
 import io.github.cdimascio.dotenv.dotenv
 
 @OptIn(PrivilegedIntent::class)
