+++ src/main/kotlin/dev/m2en/citation/command/message/RegisterCommand.kt
@@ -1,11 +1,10 @@
-package dev.m2en.citation.command
+package dev.m2en.citation.command.message
 
 import dev.kord.common.entity.Permission
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.Kord
 import dev.kord.core.behavior.edit
 import dev.kord.core.behavior.reply
-import dev.kord.core.entity.Message
 import dev.kord.core.entity.ReactionEmoji
 import dev.kord.core.event.message.MessageCreateEvent
 import dev.kord.core.kordLogger
